requirements.txt
Django==2.1.3
psycopg2==2.7.6
djangorestframework==3.9.0
gunicorn==19.9.0
